dhk/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.11.16
django-allauth==0.38.0
django-anymail==4.3
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-pipeline==1.6.14
djangorestframework==3.8.2
enum34==1.1.6
Fabric==1.14.0
futures==3.2.0
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
meld3==1.0.2
oauthlib==2.1.0
paramiko==2.4.2
pyasn1==0.4.4
pycparser==2.19
PyNaCl==1.3.0
python-openid==2.2.5
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.2.4
supervisor==3.3.4
urllib3==1.23
